logging: Re-organize code into separate modules
The new structure matches that of Python's own logging package and the
modules are more focused.  Both should be better for understanding the
framework.
